Minor edit

+++ analyzetool/auxfitting.py
@@ -1710,7 +1710,7 @@ polar-eps         1e-06
                 gas_run = subprocess.Popen("exec "+commd2, shell=True, universal_newlines='expand_cr')
                 rungas = True
             
-            simlen = (nsteps*2/1000)
+            simlen = int(nsteps*2/1000)
             commandd = ' '.join(csplit) + ' > liquid.log 2>&1'
             if elfn != 0:
                 runfn = os.path.abspath(analyzetool.run_sim.__file__)
